use messages from base and refactor token api
