+++ Cargo.toml
@@ -23,7 +23,7 @@ test-api-119 = []
 crate-type = ["lib"]
 
 [dependencies]
-reduct-base = { version = "1.18.0" }
+reduct-base = "1.19.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 reqwest = { version = "0.13", default-features = false, features = ["json", "rustls", "stream", "cookies"] }
