+++ src/client.rs
@@ -6,7 +6,6 @@
 use crate::bucket::BucketBuilder;
 use crate::http_client::HttpClient;
 use crate::replication::ReplicationBuilder;
-use crate::token::{TokenCreateOptions, TokenInfo, TokenInfoList};
 use crate::Bucket;
 use reduct_base::error::{ErrorCode, ReductError};
 use reduct_base::msg::replication_api::{
@@ -14,10 +13,13 @@ use reduct_base::msg::replication_api::{
     ReplicationSettings,
 };
 use reduct_base::msg::server_api::{BucketInfoList, ServerInfo};
-use reduct_base::msg::token_api::{Permissions, Token, TokenCreateResponse, TokenList};
 use reqwest::{Certificate, Method, Url};
 use std::sync::Arc;
 use std::time::Duration;
+
+mod token;
+
+pub use token::CreateTokenBuilder;
 
 pub struct ReductClientBuilder {
     url: String,
@@ -253,127 +255,6 @@ impl ReductClient {
         self.http_client.send_request(request).await?;
         Ok(())
     }
-    /// Get the token with permissions for the current user.
-    ///
-    /// # Returns
-    ///
-    /// The token or HttpError
-    pub async fn me(&self) -> Result<Token> {
-        self.http_client
-            .send_and_receive_json::<(), Token>(Method::GET, "/me", None)
-            .await
-    }
-
-    /// Get extended token info for the current user.
-    pub async fn me_info(&self) -> Result<TokenInfo> {
-        self.http_client
-            .send_and_receive_json::<(), TokenInfo>(Method::GET, "/me", None)
-            .await
-    }
-
-    /// Create an access token.
-    ///
-    /// Deprecated: use [`Self::create_token_with_options`] for full API support.
-    pub async fn create_token(&self, name: &str, permissions: Permissions) -> Result<String> {
-        let token = self
-            .create_token_with_options(
-                name,
-                TokenCreateOptions {
-                    permissions,
-                    ..Default::default()
-                },
-            )
-            .await?;
-        Ok(token.value)
-    }
-
-    /// Create an access token with extended options (`expires_at`, `ttl`, `ip_allowlist`).
-    pub async fn create_token_with_options(
-        &self,
-        name: &str,
-        options: TokenCreateOptions,
-    ) -> Result<TokenCreateResponse> {
-        self.http_client
-            .send_and_receive_json::<TokenCreateOptions, TokenCreateResponse>(
-                Method::POST,
-                &format!("/tokens/{}", name),
-                Some(options),
-            )
-            .await
-    }
-
-    /// Rotate an access token value and revoke the old one.
-    pub async fn rotate_token(&self, name: &str) -> Result<TokenCreateResponse> {
-        self.http_client
-            .send_and_receive_json::<(), TokenCreateResponse>(
-                Method::POST,
-                &format!("/tokens/{}/rotate", name),
-                None,
-            )
-            .await
-    }
-
-    /// Get an access token
-    ///
-    /// # Arguments
-    ///
-    /// * `name` - The name of the token
-    ///
-    /// # Returns
-    ///
-    /// The token or an error
-    pub async fn get_token(&self, name: &str) -> Result<Token> {
-        self.http_client
-            .send_and_receive_json::<(), Token>(Method::GET, &format!("/tokens/{}", name), None)
-            .await
-    }
-
-    /// Get extended access token info.
-    pub async fn get_token_info(&self, name: &str) -> Result<TokenInfo> {
-        self.http_client
-            .send_and_receive_json::<(), TokenInfo>(Method::GET, &format!("/tokens/{}", name), None)
-            .await
-    }
-
-    /// Delete an access token
-    ///
-    /// # Arguments
-    ///
-    /// * `name` - The name of the token
-    ///
-    /// # Returns
-    ///
-    /// Ok if the token was deleted, otherwise an error
-    pub async fn delete_token(&self, name: &str) -> Result<()> {
-        let request = self
-            .http_client
-            .request(Method::DELETE, &format!("/tokens/{}", name));
-        self.http_client.send_request(request).await?;
-        Ok(())
-    }
-
-    /// List all access tokens
-    ///
-    /// # Returns
-    ///
-    /// The list of tokens or an error
-    pub async fn list_tokens(&self) -> Result<Vec<Token>> {
-        let list = self
-            .http_client
-            .send_and_receive_json::<(), TokenList>(Method::GET, "/tokens", None)
-            .await?;
-        Ok(list.tokens)
-    }
-
-    /// List all access tokens with extended fields.
-    pub async fn list_tokens_info(&self) -> Result<Vec<TokenInfo>> {
-        let list = self
-            .http_client
-            .send_and_receive_json::<(), TokenInfoList>(Method::GET, "/tokens", None)
-            .await?;
-        Ok(list.tokens)
-    }
-
     /// Get list of replications
     ///
     /// # Returns
@@ -668,132 +549,6 @@ YyRIHN8wfdVoOw==
             let client = client.await;
             let bucket = client.get_bucket("test-bucket-1").await.unwrap();
             assert_eq!(bucket.name(), "test-bucket-1");
-        }
-    }
-
-    mod token_api {
-        use super::*;
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_me(#[future] client: ReductClient) {
-            let token = client.await.me().await.unwrap();
-            assert_eq!(token.name, "init-token");
-            assert!(token.permissions.unwrap().full_access);
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_me_info(#[future] client: ReductClient) {
-            let token = client.await.me_info().await.unwrap();
-            assert_eq!(token.name, "init-token");
-            assert!(token.permissions.unwrap().full_access);
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_create_token(#[future] client: ReductClient) {
-            let token_value = client
-                .await
-                .create_token(
-                    "test-token",
-                    Permissions {
-                        full_access: false,
-                        read: vec!["test-bucket".to_string()],
-                        write: vec!["test-bucket".to_string()],
-                    },
-                )
-                .await
-                .unwrap();
-
-            assert!(token_value.starts_with("test-token"));
-        }
-
-        #[cfg(feature = "test-api-119")]
-        #[rstest]
-        #[tokio::test]
-        async fn test_create_token_with_options(#[future] client: ReductClient) {
-            let token = client
-                .await
-                .create_token_with_options(
-                    "test-token-options",
-                    TokenCreateOptions {
-                        permissions: Permissions {
-                            full_access: false,
-                            read: vec!["test-bucket".to_string()],
-                            write: vec!["test-bucket".to_string()],
-                        },
-                        ttl: Some(3600),
-                        ip_allowlist: vec!["127.0.0.1".to_string()],
-                        ..Default::default()
-                    },
-                )
-                .await
-                .unwrap();
-
-            assert!(token.value.starts_with("test-token-options"));
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_get_token(#[future] client: ReductClient) {
-            let token = client.await.get_token("init-token").await.unwrap();
-            assert_eq!(token.name, "init-token");
-            assert!(token.is_provisioned);
-
-            let permissions = token.permissions.unwrap();
-            assert!(permissions.full_access);
-            assert!(permissions.read.is_empty());
-            assert!(permissions.write.is_empty());
-        }
-
-        #[cfg(feature = "test-api-119")]
-        #[rstest]
-        #[tokio::test]
-        async fn test_get_token_info(#[future] client: ReductClient) {
-            let token = client.await.get_token_info("init-token").await.unwrap();
-            assert_eq!(token.name, "init-token");
-            assert!(token.is_provisioned);
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_list_tokens(#[future] client: ReductClient) {
-            let tokens = client.await.list_tokens().await.unwrap();
-            assert!(!tokens.is_empty());
-        }
-
-        #[cfg(feature = "test-api-119")]
-        #[rstest]
-        #[tokio::test]
-        async fn test_list_tokens_info(#[future] client: ReductClient) {
-            let tokens = client.await.list_tokens_info().await.unwrap();
-            assert!(!tokens.is_empty());
-        }
-
-        #[cfg(feature = "test-api-119")]
-        #[rstest]
-        #[tokio::test]
-        async fn test_rotate_token(#[future] client: ReductClient) {
-            let client = client.await;
-            client
-                .create_token("test-token-rotate", Permissions::default())
-                .await
-                .unwrap();
-            let rotated = client.rotate_token("test-token-rotate").await.unwrap();
-            assert!(rotated.value.starts_with("test-token-rotate"));
-            client.delete_token("test-token-rotate").await.unwrap();
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn delete_token(#[future] client: ReductClient) {
-            let client = client.await;
-            client
-                .create_token("test-token", Permissions::default())
-                .await
-                .unwrap();
-            client.delete_token("test-token").await.unwrap();
         }
     }
